int_max_fix

+++ pointers_arrays_strings/100-atoi.c
@@ -6,7 +6,7 @@
 
 int _atoi(char *s)
 {
-	int num = 0;
+	long int num = 0;
 	int sign = 1;
 	int len = 0;
 	int lens;
